requirements.txt
disnake==2.4.0
asyncpg==0.25.0
python-dotenv==0.19.2
black==21.12b0
pre-commit==2.16.0
mypy==0.930
PyYAML==6.0
